// apps/web/src/components/luma/types.ts
export type LumaSemanticColor =
  | "primary"
  | "secondary"
  | "light"
  | "brand"
  | "success"
  | "error"
  | "warning"
  | "barney"
  | "blue"
  | "gray"
  | "green"
  | "orange"
  | "purple"
  | "red"
  | "yellow";

export type LumaBannerVariant = "neutral" | "success" | "warning" | "error";

export type LumaPillStyle = "solid" | "outline" | "ghost";

export type LumaImageShape = "circle" | "square" | "portrait" | "landscape";